+++ packages/orq-rc/src/models/operations/createdataset.ts
@@ -138,7 +138,7 @@ export const CreateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-09-10T14:47:45.818Z",
+    "2026-09-10T15:38:22.518Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {