+++ packages/orq-rc/RELEASES.md
@@ -12259,3 +12259,13 @@ Based on:
 - [typescript v4.15.0-rc.37] packages/orq-rc
 ### Releases
 - [NPM v4.15.0-rc.37] https://www.npmjs.com/package/@orq-ai/node/v/4.15.0-rc.37 - packages/orq-rc
+
+## 2026-09-10 15:41:42
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.796.4 (2.935.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.15.0-rc.38] packages/orq-rc
+### Releases
+- [NPM v4.15.0-rc.38] https://www.npmjs.com/package/@orq-ai/node/v/4.15.0-rc.38 - packages/orq-rc